[wiki:branches/pages] #505: lots of changes to make it use XML files as a default 	its not complete yet, wiktiext caches aren't flushed at the right times, but the admin/pages interface is much improved. 	overriding of core page properties is only done in all or nothing style. it would probably be helpful to have a revert button actually. 	page properties can be reverted etc 	main pages index list reverts instead of deletes page with xml files message alterations: 	new message type: "help" with a question mark as the icon 	messages use their own basic tables so the icons are in their own columns

+++ system/application/views/admin/pages_index.php
@@ -56,8 +56,13 @@ foreach ($pages as $page) {
 	echo('<a href="/admin/pages/page/edit/'.$escaped_codename.'">');
 	echo('edit');
 	echo('</a>');
-	if ($permissions['page_delete']) {
-		echo(', <a href="/admin/pages/page/delete/'.$escaped_codename.'">delete</a>');
+	if ($permissions['page_delete'] && $page['override']) {
+		if (!$page['default']) {
+			echo(', <a href="/admin/pages/page/delete/'.$escaped_codename.'">delete</a>');
+		}
+		else {
+			echo(', <a href="/admin/pages/page/delete/'.$escaped_codename.'">revert</a>');
+		}
 	}
 	echo(')');
 	echo('<br />');
